refactor transform golden xlsx helpers
