+++ crates/rulemorph/tests/common/mod.rs
@@ -1,3 +1,4 @@
 pub(crate) mod golden;
 pub(crate) mod trace;
 pub(crate) mod validation;
+pub(crate) mod xlsx;
